LPS-55928 move dictionary readers to suggest package
