+++ modules/portal/portal-search-lucene/src/com/liferay/portal/search/lucene/internal/LuceneSpellCheckIndexWriter.java
@@ -16,12 +16,12 @@ package com.liferay.portal.search.lucene.internal;
 
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.search.DictionaryEntry;
-import com.liferay.portal.kernel.search.DictionaryReader;
 import com.liferay.portal.kernel.search.DocumentImpl;
 import com.liferay.portal.kernel.search.SearchContext;
 import com.liferay.portal.kernel.search.SearchException;
 import com.liferay.portal.kernel.search.suggest.BaseSpellCheckIndexWriter;
+import com.liferay.portal.kernel.search.suggest.DictionaryEntry;
+import com.liferay.portal.kernel.search.suggest.DictionaryReader;
 import com.liferay.portal.kernel.search.suggest.NGramHolder;
 import com.liferay.portal.kernel.search.suggest.NGramHolderBuilderUtil;
 import com.liferay.portal.kernel.search.suggest.SuggestionConstants;
