+++ portal-service/src/com/liferay/portal/kernel/search/suggest/BaseGenericSpellCheckIndexWriter.java
@@ -16,8 +16,6 @@ package com.liferay.portal.kernel.search.suggest;
 
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.search.DictionaryEntry;
-import com.liferay.portal.kernel.search.DictionaryReader;
 import com.liferay.portal.kernel.search.Document;
 import com.liferay.portal.kernel.search.DocumentImpl;
 import com.liferay.portal.kernel.search.Field;
